move test packages
