+++ app/src/test/java/com/example/app/ExampleRoboTest.java
@@ -1,0 +1,11 @@
+package com.example.app;
+
+import org.junit.Test;
+
+public class ExampleRoboTest extends RoboExampleTestCase {
+
+    @Test
+    public void testSomething() {
+
+    }
+}
