+++ apptests/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.app.apptests.test">
+    package="com.example.app.test">
 
     <uses-sdk tools:overrideLibrary="android.support.test.uiautomator.v18" />
