+++ apptests/src/main/java/com/example/app/test/ExampleTest.java
@@ -1,0 +1,13 @@
+package com.example.app.test;
+
+import com.example.app.MainActivity_;
+
+import org.junit.Test;
+
+public class ExampleTest extends ExampleAppTestCase<MainActivity_> {
+
+    @Test
+    public void testSomething() {
+
+    }
+}
